+++ tests/config.py
@@ -30,7 +30,7 @@ class TestConfig(BaseTestClass):
             self.assertEqual(conf_data.cms_version, 3.2)
             self.assertEqual(conf_data.django_version, 1.6)
         else:
-            self.assertEqual(conf_data.cms_version, 3.2)
+            self.assertEqual(conf_data.cms_version, 3.3)
             self.assertEqual(conf_data.django_version, 1.8)
         self.assertEqual(conf_data.i18n, 'yes')
         self.assertEqual(conf_data.reversion, 'yes')
@@ -69,7 +69,7 @@ class TestConfig(BaseTestClass):
             self.assertEqual(conf_data.cms_version, 3.2)
             self.assertEqual(conf_data.django_version, 1.6)
         else:
-            self.assertEqual(conf_data.cms_version, 3.2)
+            self.assertEqual(conf_data.cms_version, 3.3)
             self.assertEqual(conf_data.django_version, 1.8)
         self.assertEqual(conf_data.i18n, 'yes')
         self.assertEqual(conf_data.reversion, 'no')
@@ -331,12 +331,12 @@ class TestConfig(BaseTestClass):
             self.assertEqual(supported_versions('beta', 'stable'), (1.9, 3.2))
             self.assertEqual(supported_versions('develop', 'stable'), (1.9, 3.2))
         else:
-            self.assertEqual(supported_versions('stable', 'stable'), (1.8, 3.2))
+            self.assertEqual(supported_versions('stable', 'stable'), (1.8, 3.3))
             self.assertEqual(supported_versions('stable', '3.1'), (1.8, 3.1))
             self.assertEqual(supported_versions('stable', '3.0.10'), (1.8, None))
-            self.assertEqual(supported_versions('stable', 'rc'), (1.8, 3.3))
-            self.assertEqual(supported_versions('stable', 'beta'), (1.8, 3.3))
-            self.assertEqual(supported_versions('stable', 'develop'), (1.8, 3.3))
+            self.assertEqual(supported_versions('stable', 'rc'), (1.8, 3.4))
+            self.assertEqual(supported_versions('stable', 'beta'), (1.8, 3.4))
+            self.assertEqual(supported_versions('stable', 'develop'), (1.8, 3.4))
 
             with self.assertRaises(RuntimeError):
                 supported_versions('stable', '2.4'), (1.5, 2.4)
@@ -344,11 +344,12 @@ class TestConfig(BaseTestClass):
                 supported_versions('1.5', 'stable'), (1.8, 3.1)
 
             self.assertEqual(supported_versions('1.5', '2.4'), (1.5, 2.4))
-            self.assertEqual(supported_versions('1.6', 'stable'), (1.6, 3.2))
-            self.assertEqual(supported_versions('1.6.9', 'stable'), (None, 3.2))
-            self.assertEqual(supported_versions('1.7', 'stable'), (1.7, 3.2))
-            self.assertEqual(supported_versions('beta', 'stable'), (1.9, 3.2))
-            self.assertEqual(supported_versions('develop', 'stable'), (1.9, 3.2))
+            with self.assertRaises(RuntimeError):
+                self.assertEqual(supported_versions('1.6', 'stable'), (1.6, 3.2))
+            self.assertEqual(supported_versions('1.6.9', 'stable'), (None, 3.3))
+            self.assertEqual(supported_versions('1.8', 'stable'), (1.8, 3.3))
+            self.assertEqual(supported_versions('beta', 'stable'), (1.9, 3.3))
+            self.assertEqual(supported_versions('develop', 'stable'), (1.9, 3.3))
 
     def test_requirements(self):
         """
@@ -421,15 +422,15 @@ class TestConfig(BaseTestClass):
             '-p'+self.project_dir,
             'example_prj'])
 
-        self.assertTrue(conf_data.requirements.find('django-cms<3.3') > -1)
+        self.assertTrue(conf_data.requirements.find('django-cms<3.4') > -1)
         if sys.version_info < (2, 7):
             self.assertTrue(conf_data.requirements.find('Django<1.7') > -1)
             self.assertTrue(conf_data.requirements.find('django-reversion>=1.8,<1.9') > -1)
         else:
             self.assertTrue(conf_data.requirements.find('Django<1.9') > -1)
             self.assertTrue(conf_data.requirements.find('django-reversion>=1.8.7') > -1)
-        self.assertTrue(conf_data.requirements.find('djangocms-text-ckeditor>=2.8') > -1)
-        self.assertTrue(conf_data.requirements.find('djangocms-admin-style>=1.0.6') > -1)
+        self.assertTrue(conf_data.requirements.find('djangocms-text-ckeditor>=3.0') > -1)
+        self.assertTrue(conf_data.requirements.find('djangocms-admin-style>=1.1.1') > -1)
         self.assertTrue(conf_data.requirements.find('djangocms-column') > -1)
         self.assertTrue(conf_data.requirements.find('djangocms-file') > -1)
         self.assertTrue(conf_data.requirements.find('djangocms-flash') == -1)